Decompilation checkpoint: 1773177428

+++ src/decomp/sas_c/newgrid_update_grid_state.c
@@ -25,7 +25,7 @@ extern LONG NEWGRID_GridStateFrameLatch;
 extern LONG NEWGRID_SelectedGridEntryPtr;
 extern LONG NEWGRID_OverridePenIndex;
 
-extern LONG NEWGRID_UpdatePresetEntry(char **entryOut, char **auxOut, WORD rowIndex, LONG keyIndex);
+extern WORD NEWGRID_UpdatePresetEntry(char **entryOut, char **auxOut, WORD rowIndex, LONG keyIndex);
 extern LONG NEWGRID2_JMPTBL_ESQ_TestBit1Based(const UBYTE *bitset, LONG bitIndex);
 extern LONG NEWGRID2_JMPTBL_DISPLIB_FindPreviousValidEntryIndex(const char *entry, const char *aux, LONG rowIndex);
 extern LONG NEWGRID_SelectEntryPen(const void *entry);
